Fixed 'prisma generate' bug

+++ src/prisma/fill/recreateRoles.ts
@@ -1,5 +1,6 @@
 import { PrismaClient, Role } from '@prisma/client';
 import * as roles from '../config/roles.json';
+import * as lib from '../../lib';
 
 const prisma = new PrismaClient();
 
@@ -8,6 +9,7 @@ const prisma = new PrismaClient();
  * and recreate from src/prisma/config/roles.json with olds id
  */
 export default async function recreateRoles(): Promise<void> {
+  lib.Console.info('Start "recreateRoles" script...');
   const oldRoles = await prisma.role.findMany();
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const rolesA: any = roles;
@@ -22,14 +24,16 @@ export default async function recreateRoles(): Promise<void> {
       lastRole = id;
       // If role not exists then create
       if (!oldRoles[id - 1]) {
+        lib.Console.info(`Creating new role: ${role.name} ...`);
         await prisma.role.create({
           data: {
             id,
             name: role.name,
           },
         });
-      } else {
+      } else if (oldRoles[id - 1].name !== role.name) {
         // Else update name
+        lib.Console.info(`Updating role: ${role.name} ...`);
         await prisma.role.update({
           where: {
             id,
@@ -41,12 +45,16 @@ export default async function recreateRoles(): Promise<void> {
       }
     }
   }
-  // Delete other which not set in config.json
-  await prisma.role.deleteMany({
-    where: {
-      id: {
-        gt: lastRole,
+  if (oldRoles.length > lastRole) {
+    // Delete other which not set in config.json
+    lib.Console.info('Deleting excesses roles ...');
+    await prisma.role.deleteMany({
+      where: {
+        id: {
+          gt: lastRole,
+        },
       },
-    },
-  });
+    });
+  }
+  lib.Console.info('Script "recreateRoles" is done!');
 }
